Debug Calendar Service: - Use getByDate() in the indicateAbsence method. - Change the primary key on Atendance to Integer instead of Long

+++ subprojects/calendar/src/main/java/nl/tudelft/sem/calendar/repositories/AttendanceRepository.java
@@ -7,10 +7,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AttendanceRepository extends JpaRepository<Attendance, Long> {
+public interface AttendanceRepository extends JpaRepository<Attendance, Integer> {
 
     List<Attendance> findAll();
-    List<Attendance> findByLectureId(long lectureId);
+    List<Attendance> findByLectureId(int lectureId);
     List<Attendance> findByStudentId(String studentId);
-    List<Attendance> findByLecureIdAndStudentId(long lectureId, String studentId);
+    List<Attendance> findByLecureIdAndStudentId(int lectureId, String studentId);
 }
